requirements.txt
python-telegram-bot==20.2
scipy==1.10.1
pillow==9.5.0
torch==2.0.0
torchvision==0.15.1 
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
